backend/requirements.txt
wheel
setuptools
black
django<4
djangorestframework
